Reland 370734: Test 'frame select -r' and fix that INT32_MIN breaks the option parser
The underlying bug was that r370734 removed the code for resetting the options. This
caused that once a 'frame select -r ...' command was typed, we kept the relative
index argument for all following 'frame select ...' invocations (even the ones with
an absolute index as they are the same command object).