[lldb][NFC] Disallow changing the ASTContext of an ClangASTContext after construction.
We have no use case in LLDB where we actually do want to change the ASTContext after
it the ClangASTContext has been constructed. All situations are just setting the
ASTContext directly after construction, so we might as well make this a Constructor
instead of supporting this tricky and unused use case.