[lldb][NFC] Create the ASTContext in ClangASTContext exactly once.
Same reason as for the previous patches about this: Having a ClangASTContext
and being able to switch the associated ASTContext isn't a use case we have
(or should have), so let's simplify all this code.

The DWARFASTParserClangTests changes are necessary as the test is using
a ClangASTContext but relied on the fact that no called function ever calls
getASTContext() on our ClangASTContext. As we now always create the ASTContext
the fact that we had an uninitialized FileSystem made the test crash.