[lldb] Pass a valid SourceLocation to Sema::RequireCompleteType in ASTResultSynthesizer
Sema::RequireCompleteTypeImpl is supposed to have an assert that checks that the
SourceLocation that is passed in is always valid. It's currently commented out but
as soon as this code goes active every LLDB expression will start crashing as
we always pass in an invalid SourceLocation.

This patch just passes in the valid SourceLocation that we have here to prevent
that from happening.