Fix an issue where the IgnoreName function was not allowing "Class" to be looked up inside a namespace or other class.
Summary: Objective C may or may not be enabled when running expressions. There was code that was no allowing "Class" or "id" to be looked up in expressions, but it was preventing these names from being able to be looked up in a namespace or class context. This fixes the issue by passing the namespace_decl along and testing it to make sure we only stop ones at the root namespace level.

Reviewers: jingham, aprantl, JDevlieghere, shafik, teemperor

Subscribers: shafik, lldb-commits, aadsm, wallace

Tags: #lldb

Differential Revision: https://reviews.llvm.org/D76964