[lldb] Don't print IRForTarget errors directly to the console
Summary:
When we get an error back from IRForTarget we directly print that error to the debugger output
stream instead of putting it in the result object. The result object only gets a vague
"The expression could not be prepared to run in the target" error message that doesn't actually
tell the user what went wrong.

This patch just puts the IRForTarget errors into the status object that is returned to the caller instead
of directly printing it to the debugger. Also updates one test that now can actually check
for the error message it is supposed to check for (instead of the default error which is all we had before).

Reviewers: JDevlieghere

Reviewed By: JDevlieghere

Differential Revision: https://reviews.llvm.org/D81654