[lldb] Add support for evaluating expressions in static member functions
Summary:
At the moment the expression parser doesn't support evaluating expressions in static member functions
and just pretends the expression is evaluated within a non-member function. This causes that all static
members are inaccessible when doing unqualified name lookup.

This patch adds support for evaluating in static member functions. It essentially just does the same setup
as what LLDB is already doing for non-static member functions (i.e., wrapping the expression in a fake
member function) with the difference that we now mark the wrapping function as static (to prevent
access to non-static members).

Reviewers: JDevlieghere, shafik

Differential Revision: https://reviews.llvm.org/D81550