[lldb] Add SubstTemplateTypeParm to RemoveWrappingTypes
Like the other type sugar removed by RemoveWrappingTypes, SubstTemplateTypeParm is just
pure sugar that should be ignored. If we don't ignore it (as we do now), LLDB will fail to
read values from record fields that have a SubstTemplateTypeParm type.

Only way to produce such a type in LLDB is to either use the `import-std-module` setting
to get a template into the expression parser or just create your own template directly
in the expression parser which is what we do in the test.

Reviewed By: jarin

Differential Revision: https://reviews.llvm.org/D85132