[lldb] Don't ignore CLEANUP ERROR results when summarizing dotest results
When we try to summarize dotest results into a single result status for lit, "CLEANUP ERROR:" are currently ignored
and we end up just returning "UNSUPPORTED". This changes the behaviour to mark the test as failed in lit.

This is pretty much the same concept as in D75031 (which was a similar bug).

Differential Revision: https://reviews.llvm.org/D83865