[lldb] Don't crash when LLDB can't extract the tsan report
Right now all tsan tests are crashing on Linux. The tests were already marked as expected failures, but since commit
20ce8affce85d added an assert that every StopInfo needs a non-empty stop description the
tests actually started crash (which is even with an expectedFailure a failed test).

The reason for that is that we never had any stop description when hitting tsan errors on Linux. Before the assert
that just made the test fail, but now the empty description is hitting the assert. This patch just adds a
generic stop description mentioning tsan to prevent that we hit that assert on platforms where we don't support
extracting the tsan report.

Reviewed By: friss

Differential Revision: https://reviews.llvm.org/D86593