[debugserver] Fix that debugserver's stop reply packets always return signal code 0
If our process terminates due to an unhandled signal, we are supposed to get
the signal code via WTERMSIG. However, we instead try to get the exit status
via WEXITSTATUS which just ends up always calculating signal code 0 (at least
on the macOS implementation where it just shifts the signal code bits away and
we're left with only 0 bits).

The exit status calculation on the LLDB side also seems a bit off as it claims
an exit status that is just the signal code (instead of for example 128 + signal code),
but that will be another patch.

Reviewed By: jasonmolenda

Differential Revision: https://reviews.llvm.org/D86336