[lldb] Set the access property on member function decls
This fixes two failures in the PDB tests.  LLVM has a "sanity" check on function decls.  One of the requirements of member functions is that they have the access property (public, protected, private) set if the function is a member function.  The check is an assert, so you'll see the failure only if you're running with assertions enabled.

This sets the access property to public to match how the existing code handles member function templates.

Reviewed By: aganea

Differential Revision: https://reviews.llvm.org/D85993