[lldb] Assert when TypeSystemClang::CreateFunctionDeclaration is called on a record
As discussed in D85993, using CreateFunctionDeclaration to add a member function to a record will probably
cause Clang to run into several assertions and is most likely not what one wants to do. All functions in a
record are CXXMethodDecls and never plain FunctionDecls. Not sure if there is an obscure Clang extension
that would support free functions inside a record, but I'm sure we don't have this use case when building an AST in LLDB.

Reviewed By: labath

Differential Revision: https://reviews.llvm.org/D86054