[ASTImporter] Add basic support for comparing Stmts and compare function bodies
Right now the ASTImporter assumes for most Expr nodes that they are always equal
which leads to non-compatible declarations ending up being merged. This patch
adds the basic framework for comparing Stmts (and with that also Exprs) and implements
the custom checks for a few Stmt subclasses. I'll implement the remaining subclasses
in follow up patches (mostly because there are a lot of subclasses and some of them
require further changes like having GNU language in the testing framework)

To make the tests possible I added that doing a structural comparison of FunctionDecls
is now also comparing the function bodies (assuming both FunctionDecls have a body).
This way the tests can just wrap any expression in a function and then we can compare
those. The alternative would be to add a separate API call to compare Stmts or to otherwise
squeeze the expressions into a type that is getting compared which seems rather hacky.

The motivation for this is that in LLDB we try to import libc++ source code and some
of the types we are importing there contain expressions (e.g. because they use `enable_if<expr>`),
so those declarations are currently merged even if they are completely different (e.g.
`enable_if<value> ...` and `enable_if<!value> ...` are currently considered equal which is
clearly not true).

Differential Revision: https://reviews.llvm.org/D87444