[lldb] Don't let dotest silently fall back to default SDK if the specified one can't be found
If the SDK name passed to dotest can't be found by `xcrun` we silently fall back to the default SDK.
This leads to rather cryptic errors being reported later on when linking the actual test executables.

Instead just directly log and abort when this situation is encountered and inform the user about
the invalid argument.

Reviewed By: JDevlieghere

Differential Revision: https://reviews.llvm.org/D89053