[lldb] Don't add no-sandbox entitlement when running tests on simulator
It seems that if codesigning the test executables with the `com.apple.private.security.no-sandbox` entitlement
then the simulator refuses to launch them and every test fails with `Process launch failed: process exited with status -1 (no such process.)`.

This patch checks if we're trying to run the test suite on the simulator and then avoids signing the executable
with `no-sandbox`.

Reviewed By: JDevlieghere

Differential Revision: https://reviews.llvm.org/D89052