[lldb][NFC] Fix some invalid escapes sequences in Python strings
I recently had to run the test suite with a debug Python which got started warning about some invalid
escape sequences in LLDB's Python code. They all attempt to add a backslash by doing a single
backslash instead of a double backslash in a normal string. This seems to work fine for now, but
Python says this behaviour is deprecated, so this patch turns all those strings into raw strings (where
a single backslash is actually a single backslash)

Reviewed By: JDevlieghere

Differential Revision: https://reviews.llvm.org/D88289