[lldb] Bump the amount of characters printed on pexpect failures from 100 to 2000
Right now when pexpect fails to find the search string in the output/buffer it just prints the last 100 characters.
Given that the usual pexpect terminal has the dimensions 24/80 (=1920 characters) this isn't enough to diagnose
bot failures with the curses UI which can repaint larger parts of the UI on every update.

This patch patches our embedded pexpect to print the last 2000 characters on failure as it seems this limit is hardcoded.

Differential Revision: https://reviews.llvm.org/D106873