[lldb] Add a test for potentially conflicting names for the Objective-C class update utility expression
We recently had an issue where a user declared a `Class::free` function which then got picked up by accident
by the expression evaluator when calling `::free`. This was due to a too lax filter in the
DWARFIndex (which was fixed by https://reviews.llvm.org/D73191 ). This broke the Objective-C utility
expression that is trying to update the Objective-C class list (which is calling `:;free`).

This adds a regression test for situations where we have a bunch of functions defined that share
the name of the global functions that this utility function calls. None of them are actually conflicting
with the global functions we are trying to call (they are all in namespaces, objects or classes).

Reviewed By: JDevlieghere

Differential Revision: https://reviews.llvm.org/D107776