[lldb] Add a test for user-defined objc_copyRealizedClassList_nolock
LLDB evaluates some utility expression to update the Objective-C class list that ends up
calling function such as `free` or `objc_copyRealizedClassList_nolock`. This adds a test that
just tries to define our own bogus version of `objc_copyRealizedClassList_nolock`. It just
tests that LLDB doesn't crash as we currently don't have a way to tell LLDB to look for
the function in a specific library.

Reviewed By: JDevlieghere

Differential Revision: https://reviews.llvm.org/D107778