[lldb][NFC] Test function-local classes and document the most obvious bugs
This feature doesn't seem to have any dedicated test. Instead some random tests
(the StructTypes and both bitfield tests) are declaring function-local classes
for no good reason. This adds a dedicated test so we can clean up the other
tests without loosing this tiny bit of test coverage.

Also add FIXME's for some basic stuff that doesn't work. The first FIXME is a good
beginner bug which just requires prepending the function name (in case we
decide to fix it instead of documenting this behaviour). The second FIXME
is caused by LLDB searching for definitions by name (which also seems to
miss the function name so there is a conflict with the outer type).