[lldb] Fix TestImportStdModule on some setups by testing minmax instead of abs
Some downstream forks of LLDB change parts of the test setup in a way that
causes lldb to somehow resolve `std::abs` (probably to `::abs`). This patches
changes the tested function here to be `std::minmax` which (hopefully) doesn't
have any identically named functions that LLDB could find. Just to be double
safe this also explicitly specified the template arguments so that in case
there is a `minmax` non-template function we still don't end up calling it
from this test.